fixed code smells and duplication
