+++ src/Altinn.Notifications.Core/Models/Orders/IBaseNotificationOrder.cs
@@ -1,6 +1,4 @@
-﻿using System.Text.Json.Serialization;
-
-using Altinn.Notifications.Core.Enums;
+﻿using Altinn.Notifications.Core.Enums;
 
 namespace Altinn.Notifications.Core.Models.Orders;
