+++ src/Altinn.Notifications.Core/Models/Orders/NotificationOrderRequest.cs
@@ -1,6 +1,4 @@
-﻿using System.Text.Json.Serialization;
-
-using Altinn.Notifications.Core.Enums;
+﻿using Altinn.Notifications.Core.Enums;
 using Altinn.Notifications.Core.Models.NotificationTemplate;
 
 namespace Altinn.Notifications.Core.Models.Orders;
