+++ src/Altinn.Notifications/Models/BaseNotificationOrderExt.cs
@@ -8,36 +8,43 @@ namespace Altinn.Notifications.Models;
 /// <remarks>
 /// External representaion to be used in the API.
 /// </remarks>
-public interface IBaseNotificationOrderExt
+public class BaseNotificationOrderExt
 {
     /// <summary>
     /// Gets or sets the id of the notification order
     /// </summary>
-    public string Id { get; set; }
-
-    /// <summary>
-    /// Gets or sets the short name of the creator of the notification order
-    /// </summary>
-    public string Creator { get; set; }
+    [JsonPropertyName("id")]
+    public string Id { get; set; } = string.Empty;
 
     /// <summary>
     /// Gets or sets the senders reference of the notification
     /// </summary>
+    [JsonPropertyName("sendersReference")]
     public string? SendersReference { get; set; }
 
     /// <summary>
     /// Gets or sets the requested send time of the notification
     /// </summary>
+    [JsonPropertyName("requestedSendTime")]
     public DateTime RequestedSendTime { get; set; }
+
+    /// <summary>
+    /// Gets or sets the short name of the creator of the notification order
+    /// </summary>
+    [JsonPropertyName("creator")]
+    public string Creator { get; set; } = string.Empty;
 
     /// <summary>
     /// Gets or sets the date and time of when the notification order was created
     /// </summary>
+    [JsonPropertyName("created")]
     public DateTime Created { get; set; }
 
     /// <summary>
     /// Gets or sets the preferred notification channel of the notification order
     /// </summary>
+    [JsonPropertyName("notificationChannel")]
+    [JsonConverter(typeof(JsonStringEnumConverter))] 
     public NotificationChannelExt NotificationChannel { get; set; }
 
     /// <summary>
