+++ src/Altinn.Notifications/Models/EmailNotificationOrderRequestExt.cs
@@ -1,5 +1,4 @@
-﻿using System.Text.Json;
-using System.Text.Json.Serialization;
+﻿using System.Text.Json.Serialization;
 
 namespace Altinn.Notifications.Models;
 
@@ -9,7 +8,7 @@ namespace Altinn.Notifications.Models;
 /// <remarks>
 /// External representation to be used in the API
 /// </remarks>
-public class EmailNotificationOrderRequestExt
+public class EmailNotificationOrderRequestExt : NotificationOrderRequestBaseExt
 {
     /// <summary>
     /// Gets or sets the subject of the email
@@ -28,42 +27,4 @@ public class EmailNotificationOrderRequestExt
     /// </summary>
     [JsonPropertyName("contentType")]
     public EmailContentTypeExt ContentType { get; set; } = EmailContentTypeExt.Plain;
-
-    /// <summary>
-    /// Gets or sets the send time of the email. Defaults to UtcNow
-    /// </summary>
-    [JsonPropertyName("requestedSendTime")]
-    public DateTime RequestedSendTime { get; set; } = DateTime.UtcNow;
-
-    /// <summary>
-    /// Gets or sets the senders reference on the notification
-    /// </summary>
-    [JsonPropertyName("sendersReference")]
-    public string? SendersReference { get; set; }
-
-    /// <summary>
-    /// Gets or sets the list of recipients
-    /// </summary>
-    [JsonPropertyName("recipients")]
-    public List<RecipientExt> Recipients { get; set; } = new List<RecipientExt>();
-
-    /// <summary>
-    /// Gets or sets whether notifications generated by this order should ignore KRR reservations
-    /// </summary>
-    [JsonPropertyName("ignoreReservation")]
-    public bool IgnoreReservation { get; set; }
-
-    /// <summary>
-    /// Gets or sets the id of the resource that the notification is related to
-    /// </summary>
-    [JsonPropertyName("resourceId")]
-    public string? ResourceId { get; set; }
-
-    /// <summary>
-    /// Json serialized the <see cref="EmailNotificationOrderRequestExt"/>
-    /// </summary>
-    public string Serialize()
-    {
-        return JsonSerializer.Serialize(this);
-    }
 }
