+++ src/Altinn.Notifications/Models/SmsNotificationOrderRequestExt.cs
@@ -9,7 +9,7 @@ namespace Altinn.Notifications.Models;
 /// <remarks>
 /// External representation to be used in the API.
 /// </remarks>
-public class SmsNotificationOrderRequestExt
+public class SmsNotificationOrderRequestExt : NotificationOrderRequestBaseExt
 {
     /// <summary>
     /// Gets or sets the sender number of the SMS 
@@ -22,42 +22,4 @@ public class SmsNotificationOrderRequestExt
     /// </summary>
     [JsonPropertyName("body")]
     public string Body { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Gets or sets the send time of the SMS. Defaults to UtcNow.
-    /// </summary>
-    [JsonPropertyName("requestedSendTime")]
-    public DateTime RequestedSendTime { get; set; } = DateTime.UtcNow;
-
-    /// <summary>
-    /// Gets or sets the senders reference on the notification
-    /// </summary>
-    [JsonPropertyName("sendersReference")]
-    public string? SendersReference { get; set; }
-
-    /// <summary>
-    /// Gets or sets the list of recipients
-    /// </summary>
-    [JsonPropertyName("recipients")]
-    public List<RecipientExt> Recipients { get; set; } = new List<RecipientExt>();
-
-    /// <summary>
-    /// Gets or sets whether notifications generated by this order should ignore KRR reservations
-    /// </summary>
-    [JsonPropertyName("ignoreReservation")]
-    public bool IgnoreReservation { get; set; }
-
-    /// <summary>
-    /// Gets or sets the id of the resource that the notification is related to
-    /// </summary>
-    [JsonPropertyName("resourceId")]
-    public string? ResourceId { get; set; }
-
-    /// <summary>
-    /// Json serialized the <see cref="SmsNotificationOrderRequestExt"/>
-    /// </summary>
-    public string Serialize()
-    {
-        return JsonSerializer.Serialize(this);
-    }
 }
