+++ test/Altinn.Notifications.Tests/Notifications.Core/TestingServices/OrderRequestServiceTests.cs
@@ -290,7 +290,7 @@ public class OrderRequestServiceTests
         Mock<IContactPointService> contactPointMock = new();
         contactPointMock
             .Setup(cp => cp.AddSmsContactPoints(It.IsAny<List<Recipient>>(), It.IsAny<string?>()))
-            .Callback<List<Recipient>>(recipients =>
+            .Callback<List<Recipient>, string?>((recipients, _) =>
             {
                 foreach (var recipient in recipients)
                 {
@@ -360,7 +360,7 @@ public class OrderRequestServiceTests
         Mock<IContactPointService> contactPointMock = new();
         contactPointMock
             .Setup(cp => cp.AddSmsContactPoints(It.IsAny<List<Recipient>>(), It.IsAny<string?>()))
-            .Callback<List<Recipient>>(recipients =>
+            .Callback<List<Recipient>, string?>((recipients, _) =>
             {
                 foreach (var recipient in recipients)
                 {
